refactor: Remove unused TD_SERVICE_HASH constant
The `TD_SERVICE_HASH` constant was defined but not used in any of the API URLs. It has been removed to clean up the constants file.

The import of `JSZip` in `services/dataService.ts` has also been removed as it is no longer needed since the logic for loading KMZ files was refactored to assume `JSZip` is available globally via CDN.